move use agents hook
